ai_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg2==2.9.11
sentence-transformers==3.0.1
torch==2.7.0
torchvision==0.22.0
Pillow==11.0.0
numpy==2.2.3
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.12
pgvector==0.3.2